Return partial template for filterable list results
This will allow the front end to use javascript to show filtered results
without having to reload the page.

When front-end js is available, it will request search results via ajax.
The front-end js will add 'partial' as a GET param to the filter url.
The back-end responds with a partial template, which the front-end will
insert into the page dynamically.