Update Wagtail admin interface for hero block
This updates the names and help_text for the fields that make up the
hero block. With this update, we have incorporated designer feedback to
bring them in line with the hero guidelines in the Design Manual, which
will make it easier for content owners to create hero graphics that fit
our site.

This uses the migration strategy laid out in our Wagtail Data Migrations
documentation for renaming Wagtail fields:
https://cfpb.github.io/cfgov-refresh/migrations/#wagtail-specific-considerations
1. Add new named fields to the block definition
2. Use a schema migration to add the new fields
3. Use a data migration to move the data from the old fields to the new
ones
4. Remove the old fields from the block definition and add another
schema migration to remove them from the schema
5. Update the block template to use the new data fields